+++ Python/NeonOcean.Debug/NeonOcean/Debug/LoggingShared.py
@@ -1,5 +1,0 @@
-import enum
-
-class LoggingModes(enum.Int):
-	Burst = 0  # type: LoggingModes
-	Continuous = 1  # type: LoggingModes